Joi is exported to the scheme folder, controllers are exported separately
