+++ controllers/getContactById.js
@@ -1,0 +1,17 @@
+const HttpError = require("../helpers");
+const contactsOperation = require("../models/contacts");
+
+const getContactById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contactsOperation.getContactById(contactId);
+    if (!result) {
+      throw HttpError(404);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getContactById;
