+++ helpers/index.js
@@ -1,4 +1,7 @@
 const HttpError = require("./HttpError");
 const withTryCatch = require("./withTryCatch");
 
-module.exports = { HttpError, withTryCatch };
+module.exports = {
+  HttpError,
+  withTryCatch,
+};
