+++ models/contacts.json
@@ -28,11 +28,5 @@
         "name": "Simon Morton",
         "email": "dui.Fusce.diam@Donec.com",
         "phone": "(233) 738-2360"
-    },
-    {
-        "id": "e437810d-51d7-4321-81be-48b98ed821ae",
-        "name": "Reuben Henryk",
-        "email": "pharetra.ut@dictum.co.uk",
-        "phone": "(715) 598-5792"
     }
 ]
