+++ routes/api/contacts.js
@@ -2,89 +2,23 @@
 const express = require("express");
 
 const router = express.Router();
-const contactsOperation = require("../../models/contacts");
-const { HttpError } = require("../../helpers");
-const Joi = require("joi");
 
-const addSchema = Joi.object({
-  name: Joi.string()
-    .required()
-    .messages({ "any.required": "Missing required name field" }),
-  email: Joi.string()
-    .required()
-    .messages({ "any.required": "Missing required email field" }),
-  phone: Joi.string()
-    .required()
-    .messages({ "any.required": "Missing required phone field" }),
-});
+const {
+  listContacts,
+  getContactById,
+  addContact,
+  updateContact,
+  removeContact,
+} = require("../../controllers");
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactsOperation.listContacts();
-    res.json(contacts);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", listContacts);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsOperation.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", getContactById);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.details[0].message);
-    }
-    const result = await contactsOperation.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", addContact);
 
-router.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, "Missing fields");
-    }
-    const { id } = req.params;
-    const result = await contactsOperation.updateContact(id, req.body);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:id", updateContact);
 
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsOperation.removeContact(id);
-    if (!result) {
-      throw HttpError(404);
-    }
-
-    res.status(200).json({
-      result,
-      message: "Contact deleted",
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:id", removeContact);
 
 module.exports = router;
